cleanup switch
